+++ torchvision/models/segmentation/__init__.py
@@ -1,3 +1,3 @@
-from .fcn import *
 from .deeplabv3 import *
+from .fcn import *
 from .lraspp import *
